Add BaseContract to allow contracts on non ActiveRecord resources
